refactor(transformer): shorten imports (#6643)
Style nit. Shorten `use` statements in transformer.
